requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.17.0
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
azure-core==1.29.1
azure-identity==1.14.0
azure-storage-blob==12.17.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.1
cryptography==41.0.3
datasets==2.14.4
decorator==5.1.1
dill==0.3.7
docker-pycreds==0.4.0
exceptiongroup==1.1.2
fastapi==0.101.0
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
h11==0.14.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.8.0
iso8601==1.1.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keybert==0.7.0
lit==16.0.6
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==9.5.0
portalocker==2.7.0
protobuf==4.23.4
psutil==5.9.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pypika-tortoise==0.1.6
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
seqeval==1.2.2
setproctitle==1.3.2
simpletransformers==0.63.11
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
starlette==0.27.0
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tortoise-orm==0.19.3
tqdm==4.65.1
transformers==4.30.2
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
uvicorn==0.23.2
validators==0.20.0
wandb==0.15.8
watchdog==3.0.0
Werkzeug==2.3.6
xxhash==3.3.0
yarl==1.9.2
zipp==3.16.2